requirements.txt
attrs==16.3.0
beautifulsoup4==4.5.1
cffi==1.9.1
constantly==15.1.0
coreapi==2.0.9
cryptography==1.6
cssselect==1.0.0
Django==1.10.3
django-rest-swagger==2.1.0
djangorestframework==3.5.3
idna==2.1
incremental==16.10.1
itypes==1.1.0
lxml==3.6.4
openapi-codec==1.1.7
parsel==1.1.0
pyasn1==0.1.9
pyasn1-modules==0.0.8
pycparser==2.17
PyDispatcher==2.0.5
pyOpenSSL==16.2.0
queuelib==1.4.2
requests==2.12.1
Scrapy==1.2.1
service-identity==16.0.0
simplejson==3.10.0
six==1.10.0
Twisted==16.5.0
uritemplate==3.0.0
w3lib==1.16.0
zope.interface==4.3.2
